requirements.txt
Flask
requests
python-decouple